Refactor: Simplify asset structure and consolidate scripts
- Merged CensorEffectBuiltin logic into the main CensorEffect script to remove redundant code and simplify the component setup.
- Restructured the asset to follow standard Unity package layout, moving all runtime scripts and resources to the root of the Runtime folder.
- Removed unnecessary subdirectories (Common, Builtin) and consolidated assembly definitions.
- Added a sample folder (Samples~) with a README to guide users on how to test the effect.
