--- Runtime/Common/CensorEffect.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 423d11adb844b3541a5aa7e9f6f767cb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
