--- Runtime/Common/ineedmypills.CensorEffect.Runtime.Common.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: db97b203b55f40146839fa847ea74817
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
