+++ Runtime/Resources/CensorEffect.shader
